Refactoring: split rename "grafanajsonsupport" package to "jsonsupport"
