+++ src/main/scala/be/cetic/tsorage/hub/grafana/jsonsupport/AnnotationResponse.scala
@@ -1,4 +1,4 @@
-package be.cetic.tsorage.hub.grafana.grafanajsonsupport
+package be.cetic.tsorage.hub.grafana.jsonsupport
 
 /**
  * A response for the annotation route ("/annotations").
