+++ src/main/scala/be/cetic/tsorage/hub/grafana/jsonsupport/SearchResponse.scala
@@ -1,4 +1,4 @@
-package be.cetic.tsorage.hub.grafana.grafanajsonsupport
+package be.cetic.tsorage.hub.grafana.jsonsupport
 
 /**
  * A response for the search route ("/search").
