+++ src/main/scala/be/cetic/tsorage/hub/grafana/jsonsupport/Target.scala
@@ -1,4 +1,4 @@
-package be.cetic.tsorage.hub.grafana.grafanajsonsupport
+package be.cetic.tsorage.hub.grafana.jsonsupport
 
 /**
  * A target (metric name).
