+++ src/main/scala/be/cetic/tsorage/hub/grafana/jsonsupport/TimeRange.scala
@@ -1,4 +1,4 @@
-package be.cetic.tsorage.hub.grafana.grafanajsonsupport
+package be.cetic.tsorage.hub.grafana.jsonsupport
 
 /**
  * A range of time in ISO 8601 date format.
